[#3820] Cleanup memberships when objects are deleted
The original bug report said that deleting an object that another object
was a member of (ex deleting a class that had a node) caused the node
page not to render.  We didn't see the rendering problem, but realized
that the membership table entries weren't getting deleted.  This fixes
that and adds some validations to the membership models.

Paired with: Nick Lewis and Igal Koshevoy