maint: More README changes, mostly formatting, based on Igal's feedback
The biggest noticeable change in the diff is unifying lines.  I don't
have a strong preference whether lines in the source markdown are single
paragraphs or fixed width, so I've made my changes consistent with the
rest of the code to be single lines.  This makes for an ugly diff in
this commit.

Fixing the information about Phusion Passenger's logs

Clearer wording on "restart webserver"

Intro sentences for sections.

More subsections for upgrading.

Backticks and other formatting to make keywords stand out.