(#4880) Fixed validation of new reports
While validation was already present, deserializing the report
during the before_validation step of assigning certain parameters
was failing, before the validation even got the chance to protect
anything. This patch adds an earlier before_validation step which
verifies that the report is actually a Puppet::Transaction::Report
and declares it invalid if not. It also logs errors and will return
406 Not Acceptable for an invalid report.