Speed up report import
Previously, we relied on ActiveRecord's nested attributes to handle
recursively storing reports for us. This is inefficient, as it does 1
insert for the report, 1 for each resource status, and 1 for each event.
For a report with 1000 resources, each with 2 events, that's ~2000
inserts.

Now we break insertion into three steps, which correspond to one insert
each. First, we create a Report instance without nested resource status
attributes, which saves the report itself, and lets us ask for its id.
We then save the resource statuses, without their events, in a single
call using the ar-extensions import method, tied to the report by id. We
then retrieve the resource statuses to get their ids, and use a map
of [type, title] => events to add the resource status id to each event.
We then store the events in a single import call. This makes one insert
for the report, one for resource statuses, and one for events.

Based on my benchmarks (using reports of 1000 resources with 2 events
each), this is a decrease in the time taken to store reports, from ~10s
per report to ~2.5s.

This commit also adds the ar-extensions gem, which is necessary for this
new implementation.